Cover radar_display, runway_overlay and wifi_setup with a recording canvas
Adds a mock LovyanGFX that records every draw call -- kind, position, colour,
order -- so layout and rendering decisions are assertable on the host. That
unlocked the three files that had zero coverage and had produced most of this
project's shipped bugs. 105 tests total, still ~3 s, still zero flash cost.

The mock mirrors the real header's global `namespace fonts` + `using namespace
fonts`, which is what made `namespace fonts = lgfx::v1::fonts;` a redeclaration
error on LovyanGFX >= 1.2.x -- so it cannot quietly accept code the device
rejects.

New coverage:
- Tag placement: no two tags overlap; every tag stays within one text line and a
  side-flip of its own symbol (the "displaced so far it looks like another
  aircraft" bug); the nearest aircraft keeps its label when space runs out; a
  dropped label never drops the aircraft.
- Dead reckoning: targets advance along track between fetches; the clamp applies
  to seen_pos + our own age, not just ours; a stale fix is dimmed rather than
  hidden; a stalled feed dims everything and then clears it entirely at expiry.
- Frame: a normal frame blits exactly once; sprite-allocation failure falls back
  to direct drawing and still reports the frame shown; the 115 KB allocation is
  not retried every frame.
- Grid: crosshairs are fillRect, not the anti-aliased wide lines that cost
  25.9 ms/frame; rings are drawn at the expected radii.
- Runways: drawn at a real airport with its ICAO label; suppressed by the
  toggle; absent mid-ocean; cached geometry stable across frames and rebuilt on
  both range and location change.
- Dataset invariant: every runway indexes an airport that exists, and the worst
  case anywhere on Earth stays inside the strip and label caps. Regenerating the
  table with medium airports would now fail loudly instead of silently
  truncating.
- BOOT button: tap latching via the ISR (including surviving a blocking fetch),
  one press = one tap, sub-debounce bounce ignored, long-press resets and
  reboots, a long hold does not also cycle the range, the range preset survives
  a reset, and the force-portal flag persists to the next boot.

Two mutations survived the first pass and were closed rather than excused:
- The strip-cap test was vacuous (LEMD has 4 strips against a cap of 32, so the
  truncation path never ran). Replaced with a real dataset invariant scan.
- The rollover test only exercised the idiom, not the code. millis() in the mock
  is now uint32_t like the device, and waitForLinkWithUi states its 32-bit
  intent explicitly, so the production function is now driven across the 49.7-day
  wrap and the unsafe `deadline = millis() + n` form fails the suite.

Verified: 105/105 host tests, firmware builds clean, hardware soak green.

+++ src/services/wifi_setup.cpp
@@ -282,8 +282,10 @@ bool waitForLinkWithUi(const char* ssid_for_ui, unsigned long attempt_ms) {
   // Elapsed-since-start, not a precomputed deadline: millis() is 32-bit and
   // wraps every 49.7 days, and `millis() < start + attempt_ms` exits instantly
   // across that boundary -- aborting every connect attempt for the duration.
-  const unsigned long started = millis();
-  while (millis() - started < attempt_ms) {
+  // uint32_t, not unsigned long: this comparison is only correct because it
+  // wraps at 2^32 exactly as millis() does. Widening it silently breaks that.
+  const uint32_t started = millis();
+  while (millis() - started < static_cast<uint32_t>(attempt_ms)) {
     if (wifiLinkUp()) {
       return true;
     }
